로또 자동 생성 클래스 분리
