+++ src/main/kotlin/lotto.kt
@@ -1,15 +1,4 @@
-class LottoGenerator {
-    fun generate(count: Int): List<List<Int>> {
-        val results = mutableListOf<List<Int>>()
-
-        repeat(count) {
-            val lotto_numbers = (1..45).shuffled().take(6).sorted()
-            results.add(lotto_numbers)
-        }
-
-        return results
-    }
-}
+package kotlin
 
 fun main() {
     var lottoCount: Int? = null
